further refactor for derby-standalone & export derby-parsing & derby-templates at top level
