+++ lib/App.server.js
@@ -16,6 +16,7 @@ var racer = require('racer');
 var util = racer.util;
 var App = require('./App');
 var files = require('./files');
+var parsing = require('../parsing');
 
 var STYLE_EXTENSIONS = ['.css'];
 var VIEW_EXTENSIONS = ['.html'];
@@ -214,10 +215,7 @@ App.prototype.deserialize = function() {
 
 App.prototype.loadViews = function(filename, namespace) {
   var data = files.loadViewsSync(this, filename, namespace);
-  for (var i = 0, len = data.views.length; i < len; i++) {
-    var item = data.views[i];
-    this.views.register(item.name, item.source, item.options);
-  }
+  parsing.registerParsedViews(this, data.views);
   if (!util.isProduction) this._watchViews(data.files, filename, namespace);
   // Make chainable
   return this;
