+++ lib/files.js
@@ -8,7 +8,7 @@ var fs = require('fs');
 var path = require('path');
 var util = require('racer/lib/util');
 var resolve = require('resolve');
-var parseViews = require('../parsing').parseViews;
+var parsing = require('../parsing');
 
 exports.loadViewsSync = loadViewsSync;
 exports.loadStylesSync = loadStylesSync;
@@ -41,20 +41,14 @@ function loadViewsSync(app, sourceFilename, namespace) {
       extensions: app.viewExtensions,
       packageFilter: deleteMain
     });
-    var extension = path.extname(importFilename);
-    var relativeNamespace = (attrs.ns == null) ?
-      path.basename(attrs.src, extension) :
-      attrs.ns;
-    var importNamespace = (namespace && relativeNamespace) ?
-      namespace + ':' + relativeNamespace :
-      namespace || relativeNamespace || '';
+    var importNamespace = parsing.getImportNamespace(namespace, attrs, importFilename);
     var imported = loadViewsSync(app, importFilename, importNamespace);
     views = views.concat(imported.views);
     files = files.concat(imported.files);
   }
 
   var htmlFile = compiler(file, filename);
-  var parsedViews = parseViews(namespace, htmlFile, filename, onImport);
+  var parsedViews = parsing.parseViews(htmlFile, namespace, filename, onImport);
   return {
     views: views.concat(parsedViews),
     files: files.concat(filename)
