+++ lib/parsing/index.js
@@ -1,6 +1,19 @@
+// TODO: Refactor and include derby-parsing module in derby itself
+exports = module.exports = require('derby-parsing');
 var htmlUtil = require('html-util');
+var path = require('path');
 
-exports.parseViews = function(namespace, file, filename, onImport) {
+exports.getImportNamespace = function(namespace, attrs, importFilename) {
+  var extension = path.extname(importFilename);
+  var relativeNamespace = (attrs.ns == null) ?
+    path.basename(attrs.src, extension) :
+    attrs.ns;
+  return (namespace && relativeNamespace) ?
+    namespace + ':' + relativeNamespace :
+    namespace || relativeNamespace || '';
+};
+
+exports.parseViews = function(file, namespace, filename, onImport) {
   var views = [];
   var prefix = (namespace) ? namespace + ':' : '';
 
@@ -51,4 +64,11 @@ exports.parseViews = function(namespace, file, filename, onImport) {
   }
 
   return views;
+};
+
+exports.registerParsedViews = function(app, items) {
+  for (var i = 0, len = items.length; i < len; i++) {
+    var item = items[i];
+    app.views.register(item.name, item.source, item.options);
+  }
 };
