+++ lib/templates/index.js
@@ -1,0 +1,2 @@
+// TODO: Refactor and include derby-templates module in derby itself
+module.exports = require('derby-templates');
